refactor: Extracted display functionality to separate module
