+++ src/display/display.h
@@ -1,0 +1,1 @@
+void displayTitle();
